hash: final fix jenkins hash p2
